test: break out integration tests
