+++ crates/node-bindings/src/nodes/mod.rs
@@ -3,16 +3,3 @@
 pub mod anvil;
 pub mod geth;
 pub mod reth;
-
-#[cfg(test)]
-mod test {
-    /// Run the given function only if we are in a CI environment.
-    pub(crate) fn ci_only<F>(f: F)
-    where
-        F: FnOnce(),
-    {
-        if ci_info::is_ci() {
-            f();
-        }
-    }
-}
